backend/run.py
#!/usr/bin/env python3
"""
Startup script for Chess Club Task Management Backend
"""

import uvicorn
import os
from dotenv import load_dotenv

def main():
    """Main function to start the FastAPI server"""
    
    # Load environment variables
    load_dotenv()
    
    # Configuration
    host = os.getenv("HOST", "0.0.0.0")
    port = int(os.getenv("PORT", "8000"))
    reload = os.getenv("RELOAD", "true").lower() == "true"
    
    print("Starting Chess Club Task Management Backend...")
    print(f"Host: {host}")
    print(f"Port: {port}")
    print(f"Reload: {reload}")
    print("=" * 50)
    
    # Start the server
    uvicorn.run(
        "main:app",
        host=host,
        port=port,
        reload=reload,
        log_level="info"
    )

if __name__ == "__main__":
    main()
