backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dateutil==2.8.2
spacy==3.7.2
dateparser==1.2.0
chess~=1.11.0
PyYAML~=6.0
requests~=2.32.0
backoff~=2.2.0
rich~=13.0.0
slowapi==0.1.9
openai==1.3.7
python-dotenv==1.0.0
aiofiles==23.2.1
psycopg2-binary
alembic
bcrypt==3.2.0
